requirements.txt
pydantic==1.10.2
python-dotenv==0.19.1
SQLAlchemy==1.4.26
psycopg2-binary>=2.9.1
anyio>=3.3.4
asgiref>=3.4.1
black>=21.9b0
certifi>=2021.10.8
charset-normalizer>=2.0.7
click>=8.0.3
dnspython>=2.1.0
email-validator>=1.1.3
h11>=0.12.0
httptools>=0.2.0
idna>=3.3
itsdangerous>=2.0.1
Jinja2>=3.0.2
MarkupSafe>=2.0.1
mypy-extensions>=0.4.3
orjson>=3.6.4
pathspec>=0.9.0
platformdirs>=2.4.0
psycopg2-binary>=2.9.1
python-multipart>=0.0.5
PyYAML>=5.4.1
regex
requests>=2.26.0
six>=1.16.0
sniffio>=1.2.0
tomli>=1.2.2
typing-extensions>=3.10.0.2
ujson>=4.2.0
urllib3>=1.26.7
uvicorn>=0.15.0
uvloop>=0.16.0
watchgod>=0.7
websockets>=10.0
nanoid==2.0.0
PyJWT==2.6.0